backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.2",
    "joblib>=1.4.2",
    "pandas>=2.2.3",
    "scikit-learn>=1.5.2",
    "google-generativeai>=0.8.3",
    "python-dotenv>=1.0.1",
    "django-cors-headers>=4.4.0",
    "djangorestframework>=3.15.2",
    "pip>=24.2",
    "setuptools>=75.1.0",
    "wheel>=0.44.0",
    "pillow>=10.4.0",
    "seaborn>=0.13.2",
]
